move DAGNodeOperation to a separate file
