+++ python/ray/dag/tests/experimental/test_execution_schedule.py
@@ -9,7 +9,7 @@ import ray.cluster_utils
 from ray.experimental.channel.torch_tensor_type import TorchTensorType
 from ray.tests.conftest import *  # noqa
 from ray.dag import InputNode, MultiOutputNode
-from ray.dag.compiled_dag_node import DAGNodeOperationType
+from ray.dag.dag_node_operation import DAGNodeOperationType
 import torch
 from typing import List
 from dataclasses import dataclass, field
